[Update] Rename LongBench v2 config suffix
