+++ opencompass/configs/datasets/longbenchv2/longbenchv2_gen.py
@@ -1,4 +1,4 @@
 from mmengine.config import read_base
 
 with read_base():
-    from .longbenchv2_gen_75fbba import LongBenchv2_datasets
+    from .longbenchv2_gen_bd9437 import LongBenchv2_datasets
